🔨 Move code to own DeleteAccountModal component
