+++ frontend/src/views/SettingsView/components/AccountSettings.vue
@@ -1,19 +1,18 @@
 <script setup lang="ts">
 import { computed, inject, ref, useTemplateRef } from 'vue';
-import { useRouter } from 'vue-router';
 import { useI18n } from 'vue-i18n';
 import { storeToRefs } from 'pinia';
 import { callKey } from '@/keys';
-import { TextInput, IconButton, ModalDialog, DangerButton, LinkButton, CheckboxInput, NoticeBar, NoticeBarTypes } from '@thunderbirdops/services-ui';
-import { PhCopySimple, PhArrowRight, PhDownloadSimple, PhX } from '@phosphor-icons/vue';
+import { TextInput, IconButton } from '@thunderbirdops/services-ui';
+import { PhCopySimple, PhArrowRight, PhDownloadSimple } from '@phosphor-icons/vue';
 import { createUserStore } from '@/stores/user-store';
 import { useSettingsStore } from '@/stores/settings-store';
-import { Alert, BlobResponse, BooleanResponse, Exception } from '@/models';
+import { BlobResponse } from '@/models';
 import { posthog, usePosthog } from '@/composables/posthog';
 import { MetricEvents } from '@/definitions';
+import DeleteAccountModal from './DeleteAccountModal.vue';
 
 const { t } = useI18n();
-const router = useRouter();
 
 const call = inject(callKey);
 const userStore = createUserStore(call);
@@ -23,10 +22,6 @@ const { currentState } = storeToRefs(settingsStore);
 const copyLinkTooltip = ref(t('label.copyLink'));
 
 const deleteModal = useTemplateRef('deleteModal');
-const consentToDeletion = ref(false);
-const confirmPassword = ref('');
-const supportUrl = import.meta.env?.VITE_SUPPORT_URL;
-const validationError = ref<Alert>(null);
 
 const displayName = computed({
   get: () => currentState.value.displayName,
@@ -44,39 +39,6 @@ const copyLink = async () => {
   setTimeout(() => {
     copyLinkTooltip.value = t('label.copyLink');
   }, 2000);
-};
-
-/**
- * Reset the form data.
- */
-const resetData = () => {
-  consentToDeletion.value = false;
-  confirmPassword.value = '';
-  validationError.value = null;
-};
-
-/**
- * Request an appointment data deletion, and then log out.
- */
-const actuallyDeleteAccount = async () => {
-  const pw = confirmPassword.value;
-
-  const { data, error }: BooleanResponse = await call('account/delete').delete({
-    password: pw,
-  }).json();
-
-  if (usePosthog) {
-    posthog.capture(MetricEvents.DeleteAccount);
-  }
-
-  if (error.value) {
-    validationError.value = { title: (data.value as Exception).detail as string };
-    return;
-  }
-
-  // We can't logout since we've deleted the user by now, so just delete local storage data.
-  userStore.$reset();
-  await router.push('/');
 };
 
 /**
@@ -144,76 +106,7 @@ const actuallyDownloadData = async () => {
     </div>
   </div>
 
-  <!-- Delete Appointment Data modal -->
-  <modal-dialog ref="deleteModal" class="delete-modal" @closed="resetData">
-    <template #header>
-      {{ t('heading.deleteAppointmentData') }}
-    </template>
-
-    <div class="delete-modal-container">
-      <notice-bar v-if="validationError" class="notice-bar" :type="NoticeBarTypes.Critical">
-        {{ validationError.title }}
-        <template #cta>
-          <icon-button @click="validationError = null" :title="t('label.close')">
-            <ph-x />
-          </icon-button>
-        </template>
-      </notice-bar>
-      
-      <p><strong>{{ t('text.settings.account.delete.permanenceHint') }}</strong></p>
-      <p>{{ t('text.settings.account.delete.impactHint') }}</p>
-      <p>{{ t('text.settings.account.delete.tbproHint') }}</p>
-      <p>
-        <checkbox-input
-          name="deletion-consent"
-          v-model="consentToDeletion"
-          :label="t('text.settings.account.delete.consent')"
-          required
-          data-testid="account-data-deletion-consent-checkbox"
-        />
-      </p>
-  
-      <div class="password-confirmation">
-        <text-input
-          name="confirm-password"
-          v-model="confirmPassword"
-          :label="t('text.settings.account.delete.confirm')"
-          type="password"
-          required
-          data-testid="account-data-deletion-confirm-password-input"
-        />
-        <danger-button
-          name="delete"
-          :disabled="!consentToDeletion || !confirmPassword"
-          @click="actuallyDeleteAccount"
-          data-testid="account-data-deletion-confirm-btn"
-        >
-          {{ t('heading.deleteAppointmentData') }}
-        </danger-button>
-      </div>
-    </div>
-
-    <template #actions>
-      <link-button
-        name="cancel"
-        @click="deleteModal?.hide()"
-        class="cancel-button"
-        data-testid="account-data-deletion-cancel-btn"
-      >
-        {{ t('label.cancel') }}
-      </link-button>
-    </template>
-
-    <template #footer>
-      <a :href="supportUrl">
-        {{ t('label.support') }}
-      </a>
-      <span>•</span>
-      <router-link to="/privacy">
-        {{ t('label.privacyPolicy') }}
-      </router-link>
-    </template>
-  </modal-dialog>
+  <delete-account-modal ref="deleteModal" />
 </template>
 
 <style scoped>
@@ -307,39 +200,6 @@ h2 {
   }
 }
 
-.delete-modal {
-  .delete-modal-container {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    padding-bottom: .25rem;
-    
-    :deep(.checkbox-control) {
-      flex-shrink: 0;
-    }
-
-    .password-confirmation {
-      display: flex;
-      flex-direction: column;
-      gap: 1rem;
-
-      label {
-        flex-grow: 0.5;
-      }
-  
-      button {
-        align-self: flex-start;
-        /* margin-top: 1.75rem; */
-        line-height: 1.25;
-      }
-    }
-  
-  }
-  .modal-actions .base.cancel-button {
-    color: var(--colour-ti-highlight);
-  }
-}
-
 @media (--md) {
   .booking-page-display-name-container {
     grid-template-columns: 20% 1fr;
@@ -356,18 +216,6 @@ h2 {
       border: 0;
       border-inline-start: 1px solid var(--colour-neutral-border);
       min-height: 53px;
-    }
-  }
-
-  .delete-modal {
-    .delete-modal-container {
-      .password-confirmation {
-        flex-direction: row;
-    
-        button {
-          margin-top: 1.75rem;
-        }
-      }
     }
   }
 }
